editoast: models: remove core_client dependency
