+++ editoast/editoast_models/src/work_schedules.rs
@@ -1,8 +1,5 @@
-use std::cmp::max;
-
 use chrono::DateTime;
 use chrono::Utc;
-use core_client::stdcm::UndirectedTrackRange;
 use editoast_derive::Model;
 use schemas::infra::TrackRange;
 use serde::Deserialize;
@@ -69,42 +66,6 @@ pub struct WorkSchedule {
     #[model(to_enum)]
     pub work_schedule_type: WorkScheduleType,
     pub work_schedule_group_id: i64,
-}
-
-impl WorkSchedule {
-    pub fn as_core_work_schedule(
-        &self,
-        earliest_departure_time: DateTime<Utc>,
-        latest_simulation_end: DateTime<Utc>,
-    ) -> Option<core_client::stdcm::WorkSchedule> {
-        let search_window_duration =
-            (latest_simulation_end - earliest_departure_time).num_milliseconds() as u64;
-
-        let start_time = elapsed_time_since_ms(&self.start_date_time, &earliest_departure_time);
-        let end_time = elapsed_time_since_ms(&self.end_date_time, &earliest_departure_time);
-
-        if end_time == 0 || start_time >= search_window_duration {
-            return None;
-        }
-
-        Some(core_client::stdcm::WorkSchedule {
-            start_time,
-            end_time,
-            track_ranges: self
-                .track_ranges
-                .iter()
-                .map(|track| UndirectedTrackRange {
-                    track_section: track.track.to_string(),
-                    begin: (track.begin * 1000.0) as u64,
-                    end: (track.end * 1000.0) as u64,
-                })
-                .collect(),
-        })
-    }
-}
-
-fn elapsed_time_since_ms(time: &DateTime<Utc>, since: &DateTime<Utc>) -> u64 {
-    max(0, (*time - since).num_milliseconds()) as u64
 }
 
 #[cfg(test)]
